-- adding confirm before employer is deleted

+++ src/employer/list.js
@@ -1,7 +1,12 @@
+import { inject } from "aurelia-framework";
 import { HttpClient } from "aurelia-http-client";
+import { DialogService } from "aurelia-dialog";
+import { Confirm } from "../dialogs/confirm";
 
+@inject(DialogService)
 export class List {
-    constructor() {
+    constructor(dialogService) {
+        this.dialogService = dialogService;
     }
 
     activate() {
@@ -21,19 +26,31 @@ export class List {
     }
 
     deleteEmployer(id) {
-        let client = new HttpClient();
-
-        client.delete(`/api/employer/${id}`).then(res => {
-            let parsedResponse = JSON.parse(res.response);
-        
-            if (parsedResponse.errors) {
-                this.apiErrors = parsedResponse.errors;
-                return;
+        let opts = {
+            viewModel: Confirm,
+            model: {
+                title: "Are you sure?",
+                message: "Are you sure you want to delete this employer?"
             }
+        };
 
-            this.apiErrors = null;
-            this.status = parsedResponse.status;            
-            this.getEmployers();
+        this.dialogService.open(opts).whenClosed(response => {
+            if (!response.wasCancelled) {        
+                let client = new HttpClient();
+
+                client.delete(`/api/employer/${id}`).then(res => {
+                    let parsedResponse = JSON.parse(res.response);
+                
+                    if (parsedResponse.errors) {
+                        this.apiErrors = parsedResponse.errors;
+                        return;
+                    }
+
+                    this.apiErrors = null;
+                    this.status = parsedResponse.status;            
+                    this.getEmployers();
+                });
+            }
         });
     }
 }
